Cache GitHub sync fingerprints to skip unchanged PRs
josh changes sync stores a JSON fingerprint (updatedAt, head OID,
comment/review counts, fetch timestamp) per change under
gh_cache/<change-id>/fingerprint on the changes ref after each full
comment fetch. On subsequent syncs, PRs whose list-query fields still
match a fresh fingerprint are skipped without fetching comments;
--no-cache forces reads to miss and --cache-ttl controls expiry.

Also restructure the GitHub change-management flow: the whole remote
sync (named sync_remote_async under block_on, GithubSyncCtx, GC,
push_pr_comments/push_pr_votes, CachePolicy, SyncStats) moves from the
sync command into josh-cli's forge layer as forge/github/changes.rs,
exposing a single sync() entrypoint; the existing forge/github.rs
auth/connection code becomes forge/github/login.rs behind a small
facade. commands/sync.rs keeps SyncArgs, a thin handle_sync
dispatcher, --clean handling, and the local/remote sync entry points.

Change: josh-sync-cache
