+++ forges/josh-github-changes/Cargo.toml
@@ -12,6 +12,7 @@ keywords = ["git", "github", "monorepo", "workflow"]
 git2.workspace = true
 anyhow.workspace = true
 url.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 
 josh-changes.workspace = true
