+++ forges/josh-github-changes/src/lib.rs
@@ -1,10 +1,12 @@
 pub mod admission;
+mod cache;
 mod comments;
 mod display;
 mod ids;
 mod prs;
 pub mod repo;
 
+pub use cache::{read_sync_fingerprint, store_sync_fingerprint, SyncFingerprint};
 pub use comments::{
     cleanup_posted_outbox_votes, fetched_comments, pending_comments, pending_votes, post_comments,
     post_votes, record_fetched_comments, PostCommentsOutcome, PostVotesOutcome, PostedComment,
