+++ forges/josh-github-graphql/src/operations/pull_request.rs
@@ -65,6 +65,9 @@ pub struct PrSummary {
     pub base_ref_oid: String,
     pub head_ref_name: String,
     pub head_oid: String,
+    pub updated_at: String,
+    pub comment_count: i64,
+    pub review_count: i64,
     pub head_commit_message: String,
     pub author_name: String,
     pub author_email: String,
@@ -204,6 +207,9 @@ impl GithubApiConnection {
                     base_ref_oid: node.base_ref_oid,
                     head_ref_name: node.head_ref_name,
                     head_oid: node.head_ref_oid,
+                    updated_at: format!("{}", node.updated_at),
+                    comment_count: node.comments.total_count,
+                    review_count: node.reviews.map(|r| r.total_count).unwrap_or(0),
                     head_commit_message: head_commit.map(|c| c.message).unwrap_or_default(),
                     author_name,
                     author_email,
